port to OSX clang with MKL

+++ marchenko3D/readTinvData3D.c
@@ -47,10 +47,10 @@ long readTinvData3D(char *filename, float *xrcv, float *yrcv, float *xsrc, float
 	fseek(fp, 0, SEEK_SET);
 	nread = fread( &hdr, 1, TRCBYTES, fp );
 	assert(nread == TRCBYTES);
-	if (hdr.scalco < 0) scl = 1.0/fabs(hdr.scalco);
+	if (hdr.scalco < 0) scl = 1.0/fabs((float)hdr.scalco);
 	else if (hdr.scalco == 0) scl = 1.0;
 	else scl = hdr.scalco;
-	if (hdr.scalel < 0) scel = 1.0/fabs(hdr.scalel);
+	if (hdr.scalel < 0) scel = 1.0/fabs((float)hdr.scalel);
 	else if (hdr.scalel == 0) scel = 1.0;
 	else scel = hdr.scalel;
 	fseek(fp, 0, SEEK_SET);
